third_party/blink/web_tests/svg/animations/svg-animation-policy-none-expected.txt

SVG with animation policy, none

This tests svg animation with animation policy none

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rootSVGElement.animationsPaused() is true
PASS rect.y.animVal.value is 0
PASS rect.y.baseVal.value is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rect.y.animVal.value is 0
PASS rect.y.baseVal.value is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rect.y.animVal.value is 0
PASS rect.y.baseVal.value is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rect.y.animVal.value is 0
PASS rect.y.baseVal.value is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rect.y.animVal.value is 0
PASS rect.y.baseVal.value is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rect.y.animVal.value is 0
PASS rect.y.baseVal.value is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.getCurrentTime() is 0
PASS rootSVGElement.getCurrentTime() is 0
PASS successfullyParsed is true

TEST COMPLETE
